requirements.txt
pydantic>=2.0.0
fastapi
uvicorn

langchain
langchain-community
langchain-groq

chromadb
sentence-transformers

pypdf
python-docx
unstructured
beautifulsoup4
requests

langchain-text-splitters
langchain-huggingface
langchain-chroma

python-multipart
python-dotenv
psycopg2-binary

pydantic_settings
alembic
SQLAlchemy